完善代码

+++ sharding-jdbc-boot-starter/springboot-sharding-jdbc-samples/springboot-sharding-jdbc-masterslave-sample/src/main/java/com/github/benyzhous/springboot/sharding/jdbc/masterslave/algorithm/SingleKeyModuloTableShardingAlgorithm.java
@@ -8,10 +8,10 @@ import com.google.common.collect.Range;
 import java.util.Collection;
 import java.util.LinkedHashSet;
 
-public class SingleKeyModuloTableShardingAlgorithm implements SingleKeyTableShardingAlgorithm<Integer> {
+public class SingleKeyModuloTableShardingAlgorithm implements SingleKeyTableShardingAlgorithm<Long> {
 
 	@Override
-	public String doEqualSharding(final Collection<String> availableTargetNames, final ShardingValue<Integer> shardingValue) {
+	public String doEqualSharding(final Collection<String> availableTargetNames, final ShardingValue<Long> shardingValue) {
 		for (String each : availableTargetNames) {
 			if (each.endsWith(shardingValue.getValue() % 4 + "")) {
 				return each;
@@ -21,10 +21,10 @@ public class SingleKeyModuloTableShardingAlgorithm implements SingleKeyTableShar
 	}
 
 	@Override
-	public Collection<String> doInSharding(final Collection<String> availableTargetNames, final ShardingValue<Integer> shardingValue) {
+	public Collection<String> doInSharding(final Collection<String> availableTargetNames, final ShardingValue<Long> shardingValue) {
 		Collection<String> result = new LinkedHashSet<>(availableTargetNames.size());
-		Collection<Integer> values = shardingValue.getValues();
-		for (Integer value : values) {
+		Collection<Long> values = shardingValue.getValues();
+		for (Long value : values) {
 			for (String each : availableTargetNames) {
 				if (each.endsWith(value % 4 + "")) {
 					result.add(each);
@@ -35,10 +35,10 @@ public class SingleKeyModuloTableShardingAlgorithm implements SingleKeyTableShar
 	}
 
 	@Override
-	public Collection<String> doBetweenSharding(final Collection<String> availableTargetNames, final ShardingValue<Integer> shardingValue) {
+	public Collection<String> doBetweenSharding(final Collection<String> availableTargetNames, final ShardingValue<Long> shardingValue) {
 		Collection<String> result = new LinkedHashSet<>(availableTargetNames.size());
-		Range<Integer> range = shardingValue.getValueRange();
-		for (Integer value = range.lowerEndpoint(); value <= range.upperEndpoint(); value++) {
+		Range<Long> range = shardingValue.getValueRange();
+		for (Long value = range.lowerEndpoint(); value <= range.upperEndpoint(); value++) {
 			for (String each : availableTargetNames) {
 				if (each.endsWith(value % 4 + "")) {
 					result.add(each);
